inspectkv: rename package `inspectkv` to `admin` (#4815)
